requirements.txt
# Microsoft Agent Framework (Preview - requires --pre flag)
agent-framework-azure-ai>=0.1.0

# Azure SDK
azure-identity>=1.15.0
azure-ai-projects>=1.0.0
azure-core>=1.29.0

# HTTP and API clients
httpx>=0.25.0
aiohttp>=3.9.0

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment management
python-dotenv>=1.0.0

# Teams Bot Framework (for Teams integration)
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0

# Utilities
typing-extensions>=4.9.0
python-dateutil>=2.8.2

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0

# Optional: For enhanced logging
structlog>=23.2.0
